refactor(threat-dialog): split add-threat dialog into tab components
