+++ apps/frontend/src/application/hooks/use-threat-measures-list.hook.ts
@@ -143,5 +143,6 @@ export const useThreatMeasuresList = ({ projectId, threatId }: UseThreatMeasures
         searchValue,
         sortBy,
         threatMeasures: sortedItems,
+        allThreatMeasures: items,
     };
 };
