+++ apps/frontend/src/utils/calcRisk.ts
@@ -3,6 +3,7 @@
  */
 
 import type { MatrixColorKey } from "#view/colors/matrix.ts";
+import type { MeasureImpact } from "#api/types/measure-impact.types.ts";
 
 /**
  * returns the correct color for a risk
@@ -51,4 +52,43 @@ export function calcRiskColour(
         return "grey";
     }
     return calcRiskColourFromRisk(damage * probability, lineOfToleranceGreen, lineOfToleranceRed);
+}
+
+/**
+ * Reduces the gross probability and damage by the given measure impacts and
+ * returns the resulting net probability, net damage and net risk.
+ *
+ * @param probability gross probability of the threat
+ * @param damage gross damage of the threat
+ * @param measureImpacts measure impacts to apply (null/undefined entries are ignored)
+ * @returns the net probability, net damage and net risk
+ */
+export function calcNetRisk(
+    probability: number,
+    damage: number,
+    measureImpacts: (MeasureImpact | null | undefined)[]
+): { netProbability: number; netDamage: number; netRisk: number } {
+    const [netProbability, netDamage] = measureImpacts.reduce<[number, number]>(
+        ([currentProbability, currentDamage], measureImpact) => {
+            if (measureImpact == null) {
+                return [currentProbability, currentDamage];
+            }
+
+            let impactedProbability = measureImpact.impactsProbability ? measureImpact.probability : currentProbability;
+            let impactedDamage = measureImpact.impactsDamage ? measureImpact.damage : currentDamage;
+
+            impactedProbability = measureImpact.setsOutOfScope ? 0 : impactedProbability;
+            impactedDamage = measureImpact.setsOutOfScope ? 0 : impactedDamage;
+
+            return [
+                impactedProbability != null && currentProbability > impactedProbability
+                    ? impactedProbability
+                    : currentProbability,
+                impactedDamage != null && currentDamage > impactedDamage ? impactedDamage : currentDamage,
+            ];
+        },
+        [probability, damage]
+    );
+
+    return { netProbability, netDamage, netRisk: netProbability * netDamage };
 }
